+++ packages/client/src/OsdkObjectFrom.test.ts
@@ -228,6 +228,7 @@ describe("Osdk", () => {
         | "__primaryKey"
         | "$objectType"
         | "$primaryKey"
+        | "$title"
       >
     >().toEqualTypeOf<
       { fooSpt: string | undefined }