+++ packages/client/src/object/convertWireToOsdkObjects/createOsdkInterface.ts
@@ -17,7 +17,7 @@
 import type { InterfaceDefinition } from "@osdk/api";
 import type { Osdk, OsdkObject } from "@osdk/client.api";
 import type { FetchedObjectTypeDefinition } from "../../ontology/OntologyProvider.js";
-import { createSimpleCache } from "../Cache.js";
+import { createSimpleCache } from "../SimpleCache.js";
 import type {
   InterfaceHolder,
   InterfaceHolderOwnProps,
@@ -72,6 +72,7 @@ function createInterfaceProxyHandler(
 
       switch (p) {
         case "$primaryKey":
+        case "$title":
         case "$objectType":
           return Reflect.getOwnPropertyDescriptor(underlying, p);
 
@@ -99,6 +100,7 @@ function createInterfaceProxyHandler(
         "$apiName",
         "$objectType",
         "$primaryKey",
+        "$title",
         ...Object.keys(newDef.properties),
       ];
     },
@@ -113,6 +115,7 @@ function createInterfaceProxyHandler(
         case "$as":
         case UnderlyingOsdkObject:
         case "$primaryKey":
+        case "$title":
         case "$objectType":
           return underlying[p as string];
       }