+++ packages/client.test.ontology/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @osdk/client.test.ontology
 
+## 1.0.0-beta.3
+
+### Minor Changes
+
+- 388dba9: Change all internal dependencies to be tilde not caret
+
+### Patch Changes
+
+- Updated dependencies [b5ac8a8]
+- Updated dependencies [388dba9]
+- Updated dependencies [700c894]
+  - @osdk/client.api@0.20.0-beta.3
+  - @osdk/api@1.9.0-beta.0
+
+## 0.1.0-beta.2
+
+### Patch Changes
+
+- Updated dependencies [141953f]
+  - @osdk/client.api@0.20.0-beta.2
+
+## 0.1.0-beta.1
+
+### Patch Changes
+
+- Updated dependencies [7adf5c7]
+- Updated dependencies [5e9d7d2]
+- Updated dependencies [67c1fd6]
+  - @osdk/client.api@0.20.0-beta.1
+
 ## 0.1.0-beta.0
 
 ### Patch Changes