+++ packages/foundry-sdk-generator/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @osdk/foundry-sdk-generator
 
+## 1.2.0-beta.2
+
+### Minor Changes
+
+- 388dba9: Change all internal dependencies to be tilde not caret
+
+### Patch Changes
+
+- Updated dependencies [388dba9]
+  - @osdk/legacy-client@2.4.0-beta.1
+  - @osdk/generator@1.12.0-beta.1
+  - @osdk/api@1.9.0-beta.0
+
+## 1.2.0-beta.1
+
+### Minor Changes
+
+- 141953f: Expose $title property in OsdkBase
+
+### Patch Changes
+
+- @osdk/legacy-client@2.4.0-beta.0
+
 ## 1.2.0-beta.0
 
 ### Patch Changes