+++ packages/maker/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @osdk/maker
 
+## 0.7.0-beta.1
+
+### Minor Changes
+
+- 388dba9: Change all internal dependencies to be tilde not caret
+- dc93aa6: Update ontology as code interface link types to reflect internal expectations of ontology integration
+
+### Patch Changes
+
+- Updated dependencies [388dba9]
+  - @osdk/api@1.9.0-beta.0
+
+## 0.7.0-beta.0
+
+### Minor Changes
+
+- 1112433: Adds support for defining interface link constraints
+- 1112433: Align syntax to be easier to read
+
 ## 0.6.0
 
 ### Minor Changes