+++ packages/shared.test/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @osdk/shared.test
 
+## 1.5.0-beta.2
+
+### Minor Changes
+
+- 388dba9: Change all internal dependencies to be tilde not caret
+
+### Patch Changes
+
+- Updated dependencies [388dba9]
+  - @osdk/api@1.9.0-beta.0
+
+## 1.5.0-beta.1
+
+### Minor Changes
+
+- 141953f: Expose $title property in OsdkBase
+
 ## 1.5.0-beta.0
 
 ### Minor Changes